increasing compatibility with foreach+curl_setopt
Just deployed a well-tested app to a vanilla PHP instance on AppFog, which is a Cloud Foundry provider. MailChimp integration went from rock solid to fully broken. After wrestling with it I found the culprit to be curl_setopt_array on line 504 here. For some reason it just wasn't setting the URL properly, causing a "No URL set!" curl error.

I know it's not your code, but this fixed everything immediately. I re-ran our tests and verified that it didn't break the local copy that was previously working, and it didn't cause any noticeable slow-down. 

Figured a quickie patch might let you guys focus on MailChimp examples instead of trying to debug this one for the next person who spins up an AppFog/Cloud Foundry instance to try things out.